backend/requirements.txt
streamlit==1.41.1
requests==2.32.3
plotly==5.24.1
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
PyJWT==2.9.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.0
